Fix recurring job not cleaning up stale drafts

+++ internet-webapp/MediaLibrary.Internet.Web/Background/ScheduledService.cs
@@ -106,24 +106,27 @@ namespace MediaLibrary.Internet.Web.Background
                     {
                         JArray imageEntities = JArray.Parse(entity.ImageEntities);
 
-                        // Delete Images
-                        for (int i = 0; i < imageEntities.Count; i++)
+                        if (imageEntities != null)
                         {
-                            var fileName = imageEntities[i]["Id"] + "_" + imageEntities[i]["Name"];
-
-                            var thumbArray = imageEntities[i]["Name"].ToString().Split(".");
-                            var thumbName = imageEntities[i]["Id"] + "_" + thumbArray[0];
-                            var middleThumbArray = thumbArray.Skip(1).Take(thumbArray.Length - 2);
-                            foreach (var thumb in middleThumbArray)
+                            // Delete Images
+                            for (int i = 0; i < imageEntities.Count; i++)
                             {
-                                thumbName += "." + thumb;
-                            }
-                            thumbName += "_thumb.jpg";
+                                var fileName = imageEntities[i]["Id"] + "_" + imageEntities[i]["Name"];
 
-                            var fileBlob = blobContainerClient.GetBlobClient(fileName);
-                            var thumnBlob = blobContainerClient.GetBlobClient(thumbName);
-                            await fileBlob.DeleteIfExistsAsync();
-                            await thumnBlob.DeleteIfExistsAsync();
+                                var thumbArray = imageEntities[i]["Name"].ToString().Split(".");
+                                var thumbName = imageEntities[i]["Id"] + "_" + thumbArray[0];
+                                var middleThumbArray = thumbArray.Skip(1).Take(thumbArray.Length - 2);
+                                foreach (var thumb in middleThumbArray)
+                                {
+                                    thumbName += "." + thumb;
+                                }
+                                thumbName += "_thumb.jpg";
+
+                                var fileBlob = blobContainerClient.GetBlobClient(fileName);
+                                var thumnBlob = blobContainerClient.GetBlobClient(thumbName);
+                                await fileBlob.DeleteIfExistsAsync();
+                                await thumnBlob.DeleteIfExistsAsync();
+                            }
                         }
 
                         // Delete Draft
